Fix Grafana plugin validation: gosec G115 + govulncheck CVEs (#200)
- gosec G115 (26 HIGH): change Granularity.Size/DateTimeFormat.Size from
  uint to int64 (idiomatic for Go's int64 time math), removing the
  overflow-prone conversions. The one remaining Pinot INT->int32 narrowing
  is 32-bit by definition and annotated with a justified #nosec.
  Verified: `gosec -severity=high` -> 0 issues.
- govulncheck (27 stdlib CVEs): bump the go directive 1.25.7 -> 1.25.11,
  which is >= every fix version on the 1.25 branch (net/url, crypto/x509,
  net/textproto, Fchmodat, ...). CI's setup-go reads this directive, so
  release binaries build with 1.25.11. Verified: a 1.25.11 binary scans
  clean with `govulncheck -mode binary`.

Staying on Go 1.25.x (not 1.26) so Snyk can process go.mod; the SDK is left
at v0.292.0 since the only newer release (v0.292.1) requires Go 1.26.3+.
The README datetimeconvert link was already fixed on main.

+++ pkg/pinot/extractors.go
@@ -101,7 +101,7 @@ func ExtractColumn(results *ResultTable, colIdx int) (any, error) {
 	case DataTypeInt:
 		return extractTypedColumn(results.RowCount(), colIdx, func(rowIdx int) (int32, error) {
 			val, err := defaultIfNil[json.Number](results.Rows[rowIdx][colIdx], json.Number("0")).Int64()
-			return int32(val), err
+			return int32(val), err // #nosec G115 -- Pinot DataTypeInt columns are 32-bit by definition; value always fits in int32
 		})
 	case DataTypeLong:
 		return extractTypedColumn(results.RowCount(), colIdx, func(rowIdx int) (int64, error) {
@@ -298,19 +298,19 @@ func ExtractColumnAsTime(results *ResultTable, colIdx int, format DateTimeFormat
 	parseLong := func(v int64) time.Time {
 		switch format.Unit {
 		case TimeUnitDays:
-			return time.Unix(86400*v*int64(format.Size), 0).UTC()
+			return time.Unix(86400*v*format.Size, 0).UTC()
 		case TimeUnitHours:
-			return time.Unix(3600*v*int64(format.Size), 0).UTC()
+			return time.Unix(3600*v*format.Size, 0).UTC()
 		case TimeUnitMinutes:
-			return time.Unix(60*v*int64(format.Size), 0).UTC()
+			return time.Unix(60*v*format.Size, 0).UTC()
 		case TimeUnitSeconds:
-			return time.Unix(v*int64(format.Size), 0).UTC()
+			return time.Unix(v*format.Size, 0).UTC()
 		case TimeUnitMilliseconds:
-			return time.UnixMilli(v * int64(format.Size)).UTC()
+			return time.UnixMilli(v * format.Size).UTC()
 		case TimeUnitMicroseconds:
-			return time.UnixMicro(v * int64(format.Size)).UTC()
+			return time.UnixMicro(v * format.Size).UTC()
 		default:
-			return time.Unix(0, v*int64(format.Size)).UTC()
+			return time.Unix(0, v*format.Size).UTC()
 		}
 	}
